fix: Resolve all compiler warnings for -Werror on g++ and clang++
- Fix sign-conversion in md5.hpp and buffer.hpp
- Fix shadow warnings in codegen templates (api.hpp.j2, endpoint.hpp.j2)
- Suppress warnings in third-party jsmn.h and reflect.hpp
- Remove unused variables and lambda captures in tests
- Clean on g++-13, g++-15, and clang++ (Apple) with -Werror

+++ include/note/backends/buffer.hpp
@@ -185,7 +185,7 @@ public:
     {
         jsmn_parser parser;
         jsmn_init(&parser);
-        token_count_ = jsmn_parse(&parser, json, len, tokens, max_tokens);
+        token_count_ = jsmn_parse(&parser, json, len, tokens, static_cast<unsigned int>(max_tokens));
         if (token_count_ < 0) token_count_ = 0;
     }
